Rename src/translate folder to klingon.
