+++ src/klingon/translate.ts
